refactor(plugin): move the wasm-pack stuff to a mixin, use logger.verbose over logger.log
